requirements.txt
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.11
django-bootstrap-v5==1.0.11
django-import-export==3.3.7
et-xmlfile==1.1.0
executing==2.0.1
Flask==3.0.3
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
MarkupPy==1.14
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
parso==0.8.4
platformdirs==4.2.1
plotly==5.21.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.2
requests==2.31.0
retrying==1.3.4
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
tablib==3.5.0
tenacity==8.2.3
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
xlrd==2.0.1
xlwt==1.3.0
zipp==3.18.1
